+++ test/IO/test_mitutoyo.py
@@ -75,6 +75,7 @@ def _compare(source, target, marker=None):
     return True
 
 
+# channel info expected from 'mitutoyo_mock.xlsx'
 EXPECTED_CHANNEL_INFO = {
     'roughness_metrics': [
         {'key': 'Ra', 'value': 0.018, 'unit': 'µm'},
@@ -84,6 +85,37 @@ EXPECTED_CHANNEL_INFO = {
         {'key': 'Rvk', 'value': 0.489, 'unit': 'µm'}],
     'cut_off': {'value': '0.08', 'unit': 'mm'},
     'acquisition_time': '2022-10-10 00:00:00',
+    'unit': 'µm'
+}
+
+# channel info expected from 'mitutoyo_2_mock.xlsx'
+EXPECTED_CHANNEL_INFO_2 = {
+    'roughness_metrics': [
+        {'key': 'Ra', 'value': 3.308, 'unit': 'µm'},
+        {'key': 'Rq', 'value': 3.959, 'unit': 'µm'},
+        {'key': 'Rz', 'value': 18.485, 'unit': 'µm'},
+        {'key': 'Rp', 'value': 8.244, 'unit': 'µm'},
+        {'key': 'Rv', 'value': 10.24, 'unit': 'µm'},
+        {'key': 'Rsk', 'value': -0.144, 'unit': ''},
+        {'key': 'Rkµ', 'value': 2.463, 'unit': ''},
+        {'key': 'Rc', 'value': 13.306, 'unit': 'µm'},
+        {'key': 'RSm', 'value': 419.9, 'unit': 'µm'},
+        {'key': 'RDq', 'value': 0.157, 'unit': ''},
+        {'key': 'Rmr', 'value': 0.03, 'unit': '%'},
+        {'key': 'Rmr(c)1', 'value': 2.911, 'unit': '%'},
+        {'key': 'Rmr(c)2', 'value': 5.822, 'unit': '%'},
+        {'key': 'Rdc', 'value': 1.828, 'unit': 'µm'},
+        {'key': 'Rt', 'value': 22.233, 'unit': 'µm'},
+        {'key': 'Rz1max', 'value': 22.062, 'unit': 'µm'},
+        {'key': 'Rk', 'value': 11.44, 'unit': 'µm'},
+        {'key': 'Rpk', 'value': 2.912, 'unit': 'µm'},
+        {'key': 'Rvk', 'value': 3.98, 'unit': 'µm'},
+        {'key': 'Mr1', 'value': 5.822, 'unit': '%'},
+        {'key': 'Mr2', 'value': 92.617, 'unit': '%'},
+        {'key': 'A1', 'value': 8.48, 'unit': ''},
+        {'key': 'A2', 'value': 14.69, 'unit': ''}],
+    'cut_off': {'value': '2.5', 'unit': 'mm'},
+    'acquisition_time': '2023-06-20 00:00:00',
     'unit': 'µm'
 }
 
@@ -119,6 +151,42 @@ def test_read_uniform(file_format_examples):
 
     # test rms roughness
     np.testing.assert_allclose(topography.rms_height_from_profile(), 0.16866328079293708)
+
+    # test for uniform flag
+    assert topography.is_uniform
+
+
+def test_read_uniform_2(file_format_examples):
+    reader = MitutoyoReader(os.path.join(file_format_examples, 'mitutoyo_2_mock.xlsx'))
+
+    assert len(reader.channels) == 1
+    assert reader.channels[0] == reader.default_channel
+
+    # test a few channel properties
+    assert reader.default_channel.unit == 'µm'
+
+    np.testing.assert_allclose(reader.default_channel.area_per_pt, 1.5)
+
+    assert reader.default_channel.dim == 1
+
+    physical_sizes = reader.default_channel.physical_sizes
+    assert len(physical_sizes) == 1
+    np.testing.assert_allclose(physical_sizes[0], 9996.)
+
+    nx, = reader.channels[0].nb_grid_pts
+    assert nx == 6664
+
+    # test channel info
+    channel_info = reader.channels[0].info
+    assert _compare(channel_info, EXPECTED_CHANNEL_INFO_2)
+
+    # test number of grid points in topography
+    topography = reader.topography()
+    nx, = topography.nb_grid_pts
+    assert nx == 6664
+
+    # test rms roughness
+    np.testing.assert_allclose(topography.rms_height_from_profile(), 3.96889742)
 
     # test for uniform flag
     assert topography.is_uniform