+++ SurfaceTopography/ChangeLog.md
@@ -7,6 +7,7 @@ v1.8.0 (not yet released)
 - API: Generalized container readers to support multiple file formats
 - API: `SurfaceContainer` is not `InMemorySurfaceContainer`
 - ENH: Reader for ZAG containers
+- MAINT: More robust date parsing for Mitutoyo reader
 
 v1.7.0 (27Jun23)
 ----------------